chore: renamed @ui5/webcomponents-api-json-parser to @ui5/webcomponents-schema-parser
