+++ apps/playground/src/app/app.component.html
@@ -1,1 +1,21 @@
-<router-outlet></router-outlet>
+
+<h3>Radio</h3>
+<section>
+  <button (click)="updateRadioFormModel()">radioControl.setValue("Option 1")</button>FormControl value: {{ radioControl.value }}, touched: {{ radioControl.touched }}<br>
+
+  <ui5-radio-button name="GroupA" text="Option 1" value="Option 1" [formControl]="radioControl"></ui5-radio-button>
+  <ui5-radio-button name="GroupA" text="Option 2" value="Option 2" [formControl]="radioControl" [checked]="true"></ui5-radio-button>
+  <ui5-radio-button name="GroupA" text="Option 3" value="Option 3" [formControl]="radioControl"></ui5-radio-button>
+</section>
+
+
+<h3>Select</h3>
+<section>
+  <button (click)="updateSelectFormModel()">selectControl.setValue("phone")</button> FormControl value: {{ selectControl.value }}, touched: {{ selectControl.touched }}<br>
+
+  <ui5-select class="select" [formControl]="selectControl">
+    <ui5-option value="phone">Phone</ui5-option>
+    <ui5-option value="tablet">Tablet</ui5-option>
+    <ui5-option value="desktop">Desktop</ui5-option>
+  </ui5-select>
+</section>
