+++ apps/playground/src/app/app.component.ts
@@ -1,11 +1,18 @@
 import {Component} from '@angular/core';
+import { FormControl } from "@angular/forms";
 
 @Component({
   selector: 'ui-angular-root',
   templateUrl: './app.component.html'
 })
 export class AppComponent {
-  log($event: any) {
-    console.log($event)
+  radioControl = new FormControl('Option 2');
+  selectControl = new FormControl('desktop');
+
+  updateSelectFormModel() {
+    this.selectControl.setValue("phone");
+  }
+  updateRadioFormModel() {
+    this.radioControl.setValue("Option 1");
   }
 }
