+++ libs/ui5-angular/wrapper.conf.ts
@@ -1,5 +1,5 @@
 import { ComponentData } from "@ui5/webcomponents-wrapper";
-import apiJsonParser from "@ui5/webcomponents-api-json-parser";
+import { apiJsonParser } from "@ui5/webcomponents-schema-parser";
 import {
   AngularExportSpecifierType,
   AngularGeneratedFile,
