+++ libs/ui5-schema-parser/.eslintrc.json
