+++ libs/ui5-schema-parser/.lib.swcrc
