+++ libs/ui5-schema-parser/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ui5/webcomponents-api-json-parser",
+  "name": "@ui5/webcomponents-schema-parser",
   "version": "0.2.3",
   "author": "SAP SE (https://www.sap.com)",
   "license": "Apache-2.0",
