+++ libs/ui5-schema-parser/project.json
@@ -1,17 +1,17 @@
 {
-  "name": "api-json-parser",
+  "name": "ui5-schema-parser",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/api-json-parser/src",
+  "sourceRoot": "libs/ui5-schema-parser/src",
   "projectType": "library",
   "targets": {
     "build": {
       "executor": "@nx/js:tsc",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/libs/api-json-parser",
-        "main": "libs/api-json-parser/src/index.ts",
-        "tsConfig": "libs/api-json-parser/tsconfig.lib.json",
-        "assets": ["libs/api-json-parser/*.md"],
+        "outputPath": "dist/libs/ui5-schema-parser",
+        "main": "libs/ui5-schema-parser/src/index.ts",
+        "tsConfig": "libs/ui5-schema-parser/tsconfig.lib.json",
+        "assets": ["libs/ui5-schema-parser/*.md"],
         "updateBuildableProjectDepsInPackageJson": true
       }
     },
@@ -19,7 +19,7 @@
       "executor": "@nx/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/api-json-parser/**/*.ts"]
+        "lintFilePatterns": ["libs/ui5-schema-parser/**/*.ts"]
       }
     }
   },
