+++ libs/ui5-schema-parser/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/api-json-parser';
