+++ libs/ui5-schema-parser/src/lib/api-json-parser/get-components.ts
