+++ libs/ui5-schema-parser/tsconfig.json
