+++ libs/ui5-schema-parser/tsconfig.lib.json
