+++ tsconfig.base.json
@@ -27,8 +27,8 @@
         "libs/library-one/library-two/src/index.ts"
       ],
       "@ui5/theming-ngx": ["libs/ui5-angular-theming/src/index.ts"],
-      "@ui5/webcomponents-api-json-parser": [
-        "libs/api-json-parser/src/index.ts"
+      "@ui5/webcomponents-schema-parser": [
+        "libs/ui5-schema-parser/src/index.ts"
       ],
       "@ui5/webcomponents-ngx": ["libs/ui5-angular/index.ts"],
       "@ui5/webcomponents-ngx-generator": [
